+++ packages/ubdev/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.11]
+### Fixed
+ - entity metadata editor show normalized metadata JSON representation ready to be stored in `meta` file
+ 
 ## [0.0.1]
 ### Changed
  - initial commit